+++ polling_stations/apps/data_collection/management/commands/import_richmond.py
@@ -3,8 +3,8 @@ Import Richmond
 """
 from django.contrib.gis.geos import Point
 
-from data_collection.management.commands import (Slugger,
-    BaseCsvStationsCsvAddressesImporter)
+from data_collection.slugger import Slugger
+from data_collection.management.commands import BaseCsvStationsCsvAddressesImporter
 from data_finder.helpers import geocode, geocode_point_only, PostcodeError
 from addressbase.models import Address
